Split duration-days e2e tests into a folder and drop its max-lines exemption
The multi-day-bookings e2e file had grown to 1,406 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split
it into test/e2e/duration-days/ — grouped by area (capacity & availability,
booking flows, admin pages, API & HTTP, and edge cases) — so every file
sits under 400 lines, then removed the file's exemption entry.

Bringing the file back under jscpd's per-file scan cap surfaced setup that
had been silently duplicated inside the monolith, including a block that
also duplicated an existing listings test. The shared pieces are now
helpers: a folder-local helpers.ts (twoGroupedListingsBookedOnAdjacentDays),
two describe-local helpers (expectNoDurationChangeLogged, groupWithSatAndCombo),
and a reusable fetchListingExportCsv in #test-utils that the existing
date-and-location export test now shares too. All original tests carry over
unchanged and pass; jscpd is back to 0% duplication.
