+++ biome.json
@@ -160,7 +160,6 @@
         "src/ui/templates/fields.ts",
         "src/ui/templates/public/reservations.tsx",
         "test/e2e/accounting.test.ts",
-        "test/e2e/duration-days.test.ts",
         "test/lib/attendee-merge.test.ts",
         "test/lib/server-attendee-form.test.ts",
         "test/lib/server-attendees.test.ts",
